pyproject.toml
[tool.poetry]
name = "run-rx"
version = "0.0.17"
description = "A tool to simplify remote execution"
authors = ["Kris Chodorow <k.chodorow@gmail.com>"]
license = "LICENSE.txt"
readme = "pip_readme.md"
homepage = "https://www.run-rx.com"
repository = "https://github.com/run-rx/rx"
packages = [
    { include = "rx" },
]
include = [
  "rx/proto/*.py",
  "rx/proto/*.pyi",
  "install/*",
]
exclude = [
  "**/test_*.py",
  "rx/proto/*.proto",
]

[tool.poetry.dependencies]
python = "^3.7"
absl-py = "^1.4.0"
grpcio = "^1.54.0"
protobuf = "^4.22.3"
PyJWT = "^2.6.0"
PyYAML = "^6.0.1"
requests = "^2.30.0"
sty = "^1.0.4"
tqdm = "^4.65.0"

[tool.poetry.scripts]
rx = 'rx.client.commands.exec:run'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
